issue #1772: wf uprava defaultnich metadat

+++ proarc-common/src/main/java/cz/cas/lib/proarc/common/process/export/mets/MetsUtils.java
@@ -1353,6 +1353,9 @@ public class MetsUtils {
             }*/
 
         }
+        if (targetFolder == null) {
+            return null;
+        }
         for (File file : targetFolder.listFiles()) {
             if (file.isDirectory()) {
                 deleteFolder(file);
